Bug #14663
In Kmelia, when the instance parameter "restrictive suppression" is enabled,
ensure that only the app manager can effectively delete a publication
and can empty the trash.

+++ kmelia/kmelia-war/src/main/webapp/kmelia/jsp/simpleListOfPublications.jsp
@@ -189,7 +189,9 @@
           operationPane.addOperation("useless", resources.getString("kmelia.operation.copyPublications"), "javascript:onclick=copyPublications()");
           operationPane.addOperation("useless", resources.getString("kmelia.operation.cutPublications"), "javascript:onclick=cutPublications()");
           operationPane.addOperation(resources.getIcon("kmelia.paste"), resources.getString("GML.paste"), "javascript:onClick=pasteFromOperations()");
-          operationPane.addOperation("useless", resources.getString("kmelia.operation.deletePublications"), "javascript:onclick=deletePublications()");
+          if (kmeliaScc.isSuppressionAllowed(profile)) {
+            operationPane.addOperation("useless", resources.getString("kmelia.operation.deletePublications"), "javascript:onclick=deletePublications()");
+          }
           operationPane.addLine();
         }
